Finish refactoring `cy.render()` to extend `Cyan`
