+++ src/index.ts
@@ -5,7 +5,7 @@ export { default as cy } from './cy'
 export { default as cleanUp } from './cleanUp'
 export { default as debug } from './debug'
 export { default as fireEvent } from './fireEvent'
-export { default as render } from './render'
+export { default as render } from './render/index'
 export { default as setupTests } from './setupTests'
 export { default as promiseQueue } from './promises/promiseQueue'
