+++ src/utils/is.utils.ts
@@ -10,20 +10,8 @@ export function isArray<T>(value: unknown): value is Array<T> {
   return Array.isArray(value)
 }
 
-export function isBool<T>(value: unknown): value is boolean {
-  return typeOf(value, 'boolean')
-}
-
-export function isBoolean<T>(value: unknown): value is boolean {
-  return typeOf(value, 'boolean')
-}
-
 export function isFunction<T>(value: unknown): value is Function {
   return typeOf(value, 'function')
-}
-
-export function isNumber<T>(value: unknown): value is number {
-  return typeOf(value, 'number')
 }
 
 export function isString<T>(value: unknown): value is string {
@@ -32,10 +20,6 @@ export function isString<T>(value: unknown): value is string {
 
 export function isObject<T>(value: unknown): value is any {
   return typeOf(value, 'object') && !isFunction(value) && !isArray(value)
-}
-
-export function isPlainObject<T>(value: unknown): value is any {
-  return Object.prototype.toString.call(value) === '[object Object]'
 }
 
 export const isHTMLCollection = (obj: unknown): boolean =>
